feat: allow `$schema` in JSON configuration file
The `$schema` property can be used by editors to provide inline
documentation and validation in JSON files (see [VS Code's
documentation](https://code.visualstudio.com/Docs/languages/json#_json-schemas-and-settings)
as an example).

Currently, the JSON schema for the `dclint` configuration file does
not allow `$schema`, so adding it causes a configuration validation
error and prevents the tool from starting.

This adds `$schema` to the configuration JSON schema and documents how
to use it.